Refactored codes for a cleaner architecture

+++ lib/components/one_ui_scroll_view/one_ui_scroll_simulation.dart
@@ -82,13 +82,13 @@ class OneUiScrollSimulation extends Simulation {
 
   @override
   double x(double time) {
-    final double t = (time / _duration).clamp(0.0, 1.0) as double;
+    final double t = (time / _duration).clamp(0.0, 1.0);
     return position + _distance * _flingDistancePenetration(t) * velocity.sign;
   }
 
   @override
   double dx(double time) {
-    final double t = (time / _duration).clamp(0.0, 1.0) as double;
+    final double t = (time / _duration).clamp(0.0, 1.0);
     return _distance * _flingVelocityPenetration(t) * velocity.sign / _duration;
   }
